move config constants
